@@ -1,29 +1,28 @@
 [build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
+requires = ["setuptools>=68.0.0", "wheel"]
+build-backend = "setuptools.build_meta"
 
 [project]
 name = "bff"
 version = "1.0.0"
 description = "Backend for Frontend - API composition layer"
-requires-python = ">=3.11"
+requires-python = ">=3.13"
 dependencies = [
-    "fastapi>=0.104.0",
+    "fastapi>=0.115.0",
     "uvicorn[standard]>=0.24.0",
-    "pydantic>=2.5.0",
+    "pydantic>=2.10.0",
     "structlog>=23.2.0",
     "httpx>=0.25.0",
-    "redis>=5.0.0",
+    "redis>=5.2.0",
     "python-multipart>=0.0.6",
     "python-dotenv>=1.0.0",
-    "opentelemetry-api>=1.21.0",
-    "opentelemetry-sdk>=1.21.0",
-    "opentelemetry-exporter-otlp-proto-grpc>=1.21.0",
-    "opentelemetry-instrumentation-fastapi>=0.42b0",
-    "opentelemetry-instrumentation-httpx>=0.42b0",
-    "opentelemetry-instrumentation-redis>=0.42b0",
-    "json-logging>=1.3.0",
-    "pythonjsonlogger>=2.0.7",
+    "opentelemetry-api>=1.28.0",
+    "opentelemetry-sdk>=1.28.0",
+    "opentelemetry-exporter-otlp-proto-grpc>=1.28.0",
+    "opentelemetry-instrumentation-fastapi>=0.49b0",
+    "opentelemetry-instrumentation-httpx>=0.49b0",
+    "opentelemetry-instrumentation-redis>=0.49b0",
+    "python-json-logger>=2.0.7",
     "pyjwt[crypto]>=2.8.0",
 ]
 
@@ -40,21 +39,25 @@ dev = [
 ]
 
 [tool.ruff]
-target-version = "py311"
+target-version = "py313"
 line-length = 120
-select = ["E", "W", "F", "I", "N", "B", "C4", "UP"]
-ignore = ["E501", "B008"]
+lint.select = ["E", "W", "F", "I", "N", "B", "C4", "UP"]
+lint.ignore = ["E501", "B008"]
 
 [tool.black]
 line-length = 120
-target-version = ['py311']
+target-version = ['py313']
 
 [tool.isort]
 profile = "black"
 line_length = 120
 
 [tool.mypy]
-python_version = "3.11"
+python_version = "3.13"
 strict = true
 warn_return_any = true
 warn_unused_configs = true
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["*"]
