@@ -1,12 +1,12 @@
-import uuid
-import time
-from typing import Optional, List, Dict, Any
-from ..ports.jwt_signer import JWTSigner
-from ...domain.services.auth_service import AuthDomainService
-from ...domain.errors import ServiceTokenError, UnauthorizedClientError
-from ...infrastructure.adapters.crypto.es256_signer import ES256Signer
-import structlog
 import os
+import uuid
+
+import structlog
+
+from domain.entities.user import ServiceToken
+from domain.errors import ServiceTokenError, UnauthorizedClientError
+from domain.services.auth_service import AuthDomainService
+from infrastructure.adapters.crypto.es256_signer import ES256Signer
 
 logger = structlog.get_logger(__name__)
 
@@ -19,13 +19,12 @@ def validate_service_client(client_id: str, client_secret: str, sub_spn: str) ->
     try:
         # Extract service name from SPN (spn:service-name -> service_name)
         service_name = sub_spn.replace("spn:", "").replace("-", "_")
-        
+
         expected_client_id = os.getenv(f"SVC_CLIENT_ID_{service_name}")
         expected_client_secret = os.getenv(f"SVC_CLIENT_SECRET_{service_name}")
-        
-        return (expected_client_id == client_id and 
-                expected_client_secret == client_secret)
-        
+
+        return expected_client_id == client_id and expected_client_secret == client_secret
+
     except Exception as e:
         logger.error("Client validation error", sub_spn=sub_spn, error=str(e))
         return False
@@ -35,13 +34,13 @@ def mint_svc_token(
     signer: ES256Signer,
     sub_spn: str,
     scopes: str,
-    actor_sub: Optional[str] = None,
-    actor_scope: Optional[str] = None,
-    actor_roles: Optional[List[str]] = None,
-    ttl: int = 300
+    actor_sub: str | None = None,
+    actor_scope: str | None = None,
+    actor_roles: list[str] | None = None,
+    ttl: int = 300,
 ) -> str:
     """Mint a service token with optional actor context"""
-    
+
     try:
         # Create JWT claims
         claims = AuthDomainService.create_service_jwt_claims(
@@ -53,42 +52,42 @@ def mint_svc_token(
             jti=str(uuid.uuid4()),
             actor_sub=actor_sub,
             actor_scope=actor_scope,
-            actor_roles=actor_roles
+            actor_roles=actor_roles,
         )
-        
+
         # Sign the token
         return signer.mint(
             sub=claims.sub,
             sid="svc",
             scopes=claims.scope or "",
             extra=claims.to_dict(),
-            ttl=ttl
+            ttl=ttl,
         )
-        
+
     except Exception as e:
         logger.error("Service token minting failed", sub_spn=sub_spn, error=str(e))
-        raise ServiceTokenError(f"Failed to mint service token: {e}")
+        raise ServiceTokenError(f"Failed to mint service token: {e}") from e
 
 
 class ServiceTokenUseCase:
     """Use case for issuing service tokens"""
-    
+
     def __init__(self, signer: ES256Signer):
         self.signer = signer
-    
+
     async def execute(
         self,
         client_id: str,
         client_secret: str,
         sub_spn: str,
         scope: str,
-        actor_sub: Optional[str] = None,
-        actor_scope: Optional[str] = None,
-        actor_roles: Optional[List[str]] = None
-    ) -> Dict[str, Any]:
+        actor_sub: str | None = None,
+        actor_scope: str | None = None,
+        actor_roles: list[str] | None = None,
+    ) -> ServiceToken:
         """
         Issue a service token
-        
+
         Args:
             client_id: Service client ID
             client_secret: Service client secret
@@ -97,23 +96,30 @@ class ServiceTokenUseCase:
             actor_sub: Acting user's subject (for on-behalf-of calls)
             actor_scope: Acting user's scopes
             actor_roles: Acting user's roles
-            
+
         Returns:
-            Dictionary with access_token, token_type, and expires_in
+            ServiceToken domain entity
         """
         try:
             # Validate client credentials
             if not validate_service_client(client_id, client_secret, sub_spn):
-                logger.warning("Invalid service client credentials", 
-                             client_id=client_id, sub_spn=sub_spn)
+                logger.warning(
+                    "Invalid service client credentials",
+                    client_id=client_id,
+                    sub_spn=sub_spn,
+                )
                 raise UnauthorizedClientError("Invalid client credentials")
-            
+
             # Get TTL from environment
             ttl = int(os.getenv("SVC_TOKEN_TTL_SECONDS", "300"))
-            
-            logger.info("Minting service token", 
-                       sub_spn=sub_spn, scope=scope, actor_sub=actor_sub)
-            
+
+            logger.info(
+                "Minting service token",
+                sub_spn=sub_spn,
+                scope=scope,
+                actor_sub=actor_sub,
+            )
+
             # Mint the token
             access_token = mint_svc_token(
                 signer=self.signer,
@@ -122,20 +128,22 @@ class ServiceTokenUseCase:
                 actor_sub=actor_sub,
                 actor_scope=actor_scope,
                 actor_roles=actor_roles,
-                ttl=ttl
+                ttl=ttl,
             )
-            
+
             logger.info("Service token minted successfully", sub_spn=sub_spn)
-            
-            return {
-                "access_token": access_token,
-                "token_type": "Bearer",
-                "expires_in": ttl
-            }
-            
+
+            return ServiceToken(
+                access_token=access_token,
+                token_type="Bearer",
+                expires_in=ttl,
+                scope=scope,
+                sub_spn=sub_spn,
+                actor_sub=actor_sub,
+            )
+
         except UnauthorizedClientError:
             raise
         except Exception as e:
-            logger.error("Service token issuance failed", 
-                        sub_spn=sub_spn, error=str(e))
-            raise ServiceTokenError(f"Failed to issue service token: {e}")
+            logger.error("Service token issuance failed", sub_spn=sub_spn, error=str(e))
+            raise ServiceTokenError(f"Failed to issue service token: {e}") from e
