@@ -1,10 +1,8 @@
-from fastapi import APIRouter, HTTPException, Request, Depends
-from pydantic import BaseModel
-from typing import Optional, List
 import structlog
+from fastapi import APIRouter, Depends, HTTPException, Request
 
-from ...application.use_cases.svc_token import ServiceTokenUseCase
-from ..schema.svc_token_schemas import ServiceTokenRequest, ServiceTokenResponse
+from application.use_cases.svc_token import ServiceTokenUseCase
+from presentation.schema.svc_token_schemas import ServiceTokenRequest, ServiceTokenResponse
 
 router = APIRouter()
 logger = structlog.get_logger(__name__)
@@ -18,7 +16,7 @@ def get_service_token_use_case(request: Request) -> ServiceTokenUseCase:
 @router.post("/svc/token", response_model=ServiceTokenResponse)
 async def create_service_token(
     request: ServiceTokenRequest,
-    svc_token_uc: ServiceTokenUseCase = Depends(get_service_token_use_case)
+    svc_token_uc: ServiceTokenUseCase = Depends(get_service_token_use_case),
 ):
     """Create a service token for service-to-service authentication"""
     try:
@@ -29,17 +27,16 @@ async def create_service_token(
             scope=request.scope,
             actor_sub=request.actor_sub,
             actor_scope=request.actor_scope,
-            actor_roles=request.actor_roles
+            actor_roles=request.actor_roles,
         )
-        
+
         logger.info("Service token issued", sub_spn=request.sub_spn)
         return ServiceTokenResponse(**result)
-        
+
     except Exception as e:
-        logger.error("Service token creation failed", 
-                    sub_spn=request.sub_spn, error=str(e))
-        
+        logger.error("Service token creation failed", sub_spn=request.sub_spn, error=str(e))
+
         if "Invalid client" in str(e) or "unauthorized" in str(e).lower():
-            raise HTTPException(status_code=401, detail="Invalid client credentials")
-        
-        raise HTTPException(status_code=500, detail="Service token creation failed")
+            raise HTTPException(status_code=401, detail="Invalid client credentials") from e
+
+        raise HTTPException(status_code=500, detail="Service token creation failed") from e
